requirements-deploy.txt
# Cloud deployment requirements with full DataBench and Plug & Play functionality
Flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.6
python-dotenv>=1.0.0

# Production server with WebSocket support
gunicorn>=21.2.0
eventlet>=0.33.3

# Authentication dependencies (optional - will gracefully fail if not available)
PyJWT>=2.8.0
bcrypt>=4.1.2

# Firebase dependencies (optional)
firebase-admin>=6.4.0

# DataBench dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.12.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0
scipy>=1.10.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
clip-by-openai>=1.0
huggingface-hub>=0.16.0

# Plug & Play dependencies
pyserial>=3.5
psutil>=5.9.0
requests>=2.31.0

# Basic utilities
PyYAML>=6.0.1
Pillow>=10.0.0
pandas>=2.0.0

# Cloud deployment essentials